wrong libft... from alex lol

+++ libft/array/array_clear.c
@@ -6,7 +6,7 @@
 /*   By: aamadori <aamadori@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/20 15:06:36 by aamadori          #+#    #+#             */
-/*   Updated: 2019/04/01 19:46:28 by aamadori         ###   ########.fr       */
+/*   Updated: 2019/01/30 12:40:26 by aamadori         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,10 @@ void	array_clear(t_array *array, void (*destroy_func)(void*))
 	size_t	index;
 
 	index = 0;
-	while (destroy_func != NULL && index < array->length)
+	while (index < array->length)
 	{
-		destroy_func(array->ptr + (index * array->elem_size));
+		if (destroy_func != NULL)
+			destroy_func(array->ptr + (index * array->elem_size));
 		index++;
 	}
 	free(array->ptr);
